+++ src/ReactPixiFiber.js
@@ -1,8 +1,8 @@
 import ReactFiberReconciler from "react-reconciler";
 import emptyObject from "fbjs/lib/emptyObject";
 import invariant from "fbjs/lib/invariant";
-import now from "performance-now";
 import {
+  unstable_now as now,
   unstable_scheduleCallback as scheduleDeferredCallback,
   unstable_cancelCallback as cancelDeferredCallback,
 } from "scheduler";