+++ src/Stage/index.js
@@ -2,6 +2,6 @@ import { areReactHooksAvailable } from "../compat";
 import createStageFunction from "./hooks";
 import createStageClass from "./legacy";
 
-export default (areReactHooksAvailable() ? createStageFunction() : createStageClass());
+export default areReactHooksAvailable() ? createStageFunction() : createStageClass();
 
 export { createStageFunction, createStageClass };